// src/types/partials/ValueSetExpansion.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TUri } from '../simpleTypes/Uri';
import { TDateTime } from '../simpleTypes/DateTime';
import { TInt } from '../simpleTypes/Int';
import { TValueSetParameter } from '../partials/ValueSetParameter';
import { TValueSetContains } from '../partials/ValueSetContains';

export type TValueSetExpansion = {
    id?: String;
    extension?: TExtension[];
    modifierExtension?: TExtension[];
    identifier?: TUri;
    timestamp: TDateTime;
    total?: TInt;
    offset?: TInt;
    parameter?: TValueSetParameter[];
    contains?: TValueSetContains[];
};